#  Allow currying of metric vec's