@beorn7 2017-12-22
Originally, I planned to have a "curry free" implementation under the hood for performance reasons, but the performance impact without currying is minimal. In particular, the whole curry handling is alloc-free.

Before:
```
BenchmarkCounterWithLabelValues-4                                       10000000               115 ns/op               0 B/op          0 allocs/op
BenchmarkCounterWithLabelValuesConcurrent-4                             20000000                94.0 ns/op             0 B/op          0 allocs/op
BenchmarkGaugeWithLabelValues-4                                         20000000               112 ns/op               0 B/op          0 allocs/op
BenchmarkSummaryWithLabelValues-4                                        2000000               786 ns/op               0 B/op          0 allocs/op
BenchmarkHistogramWithLabelValues-4                                     10000000               143 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValuesBasic-4                                10000000               140 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues2Keys10ValueCardinality-4              10000000               153 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues4Keys10ValueCardinality-4               5000000               251 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues2Keys100ValueCardinality-4             10000000               164 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues10Keys100ValueCardinality-4             2000000               607 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues10Keys1000ValueCardinality-4            2000000               689 ns/op               0 B/op          0 allocs/op
```

After:
```
BenchmarkCounterWithLabelValues-4                                       10000000               127 ns/op               0 B/op          0 allocs/op
BenchmarkCounterWithLabelValuesConcurrent-4                             20000000                99.8 ns/op             0 B/op          0 allocs/op
BenchmarkGaugeWithLabelValues-4                                         10000000               121 ns/op               0 B/op          0 allocs/op
BenchmarkSummaryWithLabelValues-4                                        2000000               772 ns/op               0 B/op          0 allocs/op
BenchmarkHistogramWithLabelValues-4                                     10000000               148 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValuesBasic-4                                10000000               146 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues2Keys10ValueCardinality-4              10000000               162 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues4Keys10ValueCardinality-4               5000000               265 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues2Keys100ValueCardinality-4             10000000               166 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues10Keys100ValueCardinality-4             2000000               617 ns/op               0 B/op          0 allocs/op
BenchmarkMetricVecWithLabelValues10Keys1000ValueCardinality-4            2000000               662 ns/op               0 B/op          0 allocs/op
```