@beorn7 2017-12-22
> I've wondered about is if users mix maps and lists, that could get confusing. It might be an idea to advise against doing do.

The order stays the same. Curried labels just disappear from it. We already advise against using the variadic functions because of ordering concerns: “Note that for more than one label value, this method is prone to mistakes caused by an incorrect order of arguments. Consider GetMetricWith(Labels) as an alternative to avoid that type of mistake.”