@beorn7 2017-12-22
I have clarified the doc comment about the ordering of remaining labels.