@dominikschulz 2017-12-22
Thanks a lot, this looks like a very good approach that could help us. @zwopir