@beorn7 2017-12-24
I still need to add some tests for the `promhttp` middleware interaction.